Readme file Completed
Crime Information Management System is an application which interact with crime and criminal database and take the information from the database and give it to the user. It is consoled based application which give functionalities to user like register crime, register criminal, show status of crime case and get report on no of solved and not solved cases. This backend application is developed using JAVA and to get the data from database I used JDBC.